refactor: NotificationController와 NotificationSettingController 분리
